api/requirements.txt
falcon
gunicorn
joblib
kafka-python
